% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_Membership.R
\name{as.membership}
\alias{as.membership}
\title{As Membership/Membership Matrix}
\usage{
as.membership(member)
}
\arguments{
\item{member}{membership matrix}
}
\value{
Membership Membership object
}
\description{
Convert matrix to membership matrix
}
\section{Slots}{

\describe{
\item{\code{member}}{membership matrix}

\item{\code{hard.label}}{hard label}
}}

