% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co.vote.R
\name{co.vote}
\alias{co.vote}
\title{Consensus / Ensemble Fuzzy Clustering: Voting}
\usage{
co.vote(object, method)
}
\arguments{
\item{object}{a fuzzycluster_list object}

\item{method}{voting step that used to combine the partition}
}
\description{
Provide consensus / ensemble fuzzy clustering with voting method. Several option for voting step provided
}

