% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_co_fuzzycluster.R
\docType{class}
\name{co_fuzzycluster-class}
\alias{co_fuzzycluster-class}
\title{Consensus Fuzzy Cluster  Result}
\description{
Consensus Fuzzy Cluster  Result
}
\section{Slots}{

\describe{
\item{\code{partition.ensemble}}{membership matrix}

\item{\code{label.ensemble}}{vector of hard partition}

\item{\code{method.consensus}}{method of fuzzy clustering used}
}}

