% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_co_fuzzycluster.R
\docType{methods}
\name{is.na,co_fuzzycluster-method}
\alias{co_fuzzycluster-method}
\alias{is.na,co_fuzzycluster-method}
\alias{method.consensus}
\alias{method.consensus,}
\alias{method.consensus,co_fuzzycluster-method}
\alias{show,co_fuzzycluster-method}
\title{Method for co_fuzzycluster classes}
\usage{
\S4method{is.na}{co_fuzzycluster}(x)

\S4method{show}{co_fuzzycluster}(object)

method.consensus(x)

\S4method{method.consensus}{co_fuzzycluster}(x)
}
\arguments{
\item{x}{an co_fuzzycluster object}

\item{object}{an co_fuzzycluster object}
}
\description{
Method for co_fuzzycluster classes
}

