% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobainher.R
\docType{class}
\name{ddddd-class}
\alias{ddddd-class}
\title{dddddd Result}
\description{
dddddd Result
}
\section{Slots}{

\describe{
\item{\code{centroid}}{centroid matrix}

\item{\code{distance}}{distance matrix}

\item{\code{func.obj}}{function objective}

\item{\code{call.func}}{called function}

\item{\code{fuzzyfier}}{fuzzyness parameter}

\item{\code{method.fuzzy}}{method of fuzzy clustering used}
}}

