% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpair_fuzzy.R
\name{fuzzy.cPair}
\alias{fuzzy.cPair}
\title{Pairing fuzzy cluster}
\usage{
fuzzy.cPair(x, y)
}
\arguments{
\item{x}{an fuzzycluster object}

\item{y}{an fuzzycluster object}
}
\description{
Pairing fuzzy cluster
}

