% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_fuzzycluster.R
\docType{class}
\name{fuzzycluster-class}
\alias{fuzzycluster-class}
\title{Fuzzy Result}
\description{
Fuzzy Result
}
\section{Slots}{

\describe{
\item{\code{partition}}{membership matrix}

\item{\code{label}}{vector of hard partition}

\item{\code{centroid}}{centroid matrix}

\item{\code{distance}}{distance matrix}

\item{\code{func.obj}}{function objective}

\item{\code{call.func}}{called function}

\item{\code{fuzzyfier}}{fuzzyness parameter}

\item{\code{method.fuzzy}}{method of fuzzy clustering used}
}}

