% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_fuzzycluster.R
\docType{methods}
\name{is.na,fuzzycluster-method}
\alias{.cPair}
\alias{.cPair,ANY,ANY-method}
\alias{.cPair,ANY,fuzzycluster-method}
\alias{.cPair,fuzzycluster,ANY-method}
\alias{.cPair,fuzzycluster,fuzzycluster-method}
\alias{cPair,fuzzycluster-method}
\alias{cPair,fuzzycluster-method,ANY-method}
\alias{call.func}
\alias{call.func,}
\alias{call.func,fuzzycluster-method}
\alias{centroid}
\alias{centroid,}
\alias{centroid,fuzzycluster-method}
\alias{distance}
\alias{distance,}
\alias{distance,fuzzycluster-method}
\alias{func.obj}
\alias{func.obj,}
\alias{func.obj,fuzzycluster-method}
\alias{fuzzycluster-method}
\alias{fuzzyfier}
\alias{fuzzyfier,}
\alias{fuzzyfier,fuzzycluster-method}
\alias{is.na,fuzzycluster-method}
\alias{label}
\alias{label,}
\alias{label,fuzzycluster-method}
\alias{method.fuzzy}
\alias{method.fuzzy,}
\alias{method.fuzzy,fuzzycluster-method}
\alias{partition}
\alias{partition,}
\alias{partition,fuzzycluster-method}
\alias{show,fuzzycluster-method}
\title{Method for fuzzycluster classes}
\usage{
\S4method{is.na}{fuzzycluster}(x)

\S4method{show}{fuzzycluster}(object)

label(x)

\S4method{label}{fuzzycluster}(x)

centroid(x)

\S4method{centroid}{fuzzycluster}(x)

label(x)

\S4method{label}{fuzzycluster}(x)

distance(x)

\S4method{distance}{fuzzycluster}(x)

func.obj(x)

\S4method{func.obj}{fuzzycluster}(x)

call.func(x)

\S4method{call.func}{fuzzycluster}(x)

fuzzyfier(x)

\S4method{fuzzyfier}{fuzzycluster}(x)

method.fuzzy(x)

\S4method{method.fuzzy}{fuzzycluster}(x)

partition(x)

\S4method{partition}{fuzzycluster}(x)

.cPair(x, y)

\S4method{.cPair}{fuzzycluster,fuzzycluster}(x, y)

\S4method{.cPair}{fuzzycluster,ANY}(x, y)

\S4method{.cPair}{ANY,fuzzycluster}(x, y)

\S4method{.cPair}{ANY,ANY}(x, y)
}
\arguments{
\item{x}{an fuzzycluster object}

\item{object}{an fuzzycluster object}

\item{y}{an fuzzycluster object}
}
\description{
Method for fuzzycluster classes
}

