% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_fuzzycluster_list.R
\docType{class}
\name{fuzzycluster_list-class}
\alias{fuzzycluster_list-class}
\title{Fuzzy Result List}
\description{
Fuzzy Result List
}
\section{Slots}{

\describe{
\item{\code{pair}}{list of fuzzy cluster object}
}}

