% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_fuzzycluster_list.R
\docType{methods}
\name{pair}
\alias{pair}
\alias{pair,fuzzycluster_list-method}
\title{Method for fuzzycluster_list classes}
\usage{
pair(object)

\S4method{pair}{fuzzycluster_list}(object)
}
\arguments{
\item{object}{an fuzzycluster_list object}
}
\description{
Method for fuzzycluster_list classes
}

