% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_fuzzyResult.R
\name{fuzzyResult-class}
\alias{fuzzyResult-class}
\alias{fuzzyResult-method}
\alias{initFuzzyResult}
\alias{partition,}
\title{Fuzzy Result}
\usage{
initFuzzyResult()
}
\description{
Fuzzy Result
}
\section{Slots}{

\describe{
\item{\code{partition}}{membership matrix}

\item{\code{label}}{vector of hard partition}

\item{\code{centroid}}{centroid matrix}

\item{\code{distance}}{distance matrix}

\item{\code{func.obj}}{function objective}

\item{\code{call.func}}{called function}

\item{\code{fuzzyfier}}{fuzzyness parameter}

\item{\code{method.fuzzy}}{method of fuzzy clustering used}
}}

