% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_Membership.R
\name{is.membership}
\alias{is.membership}
\title{Check membership Matrix/Object}
\usage{
is.membership(object)
}
\arguments{
\item{object}{an object that used for membership checking}
}
\description{
checking object is membership object or not
}

