% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_membership.R
\docType{class}
\name{membership-class}
\alias{membership-class}
\title{Membership}
\description{
Membership
}
\section{Slots}{

\describe{
\item{\code{member}}{membership matrix}

\item{\code{hard.label}}{vector of hard hard.labeling}
}}

