% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_membership.R
\docType{methods}
\name{is.na,membership-method}
\alias{hard.label}
\alias{hard.label,membership-method}
\alias{is.na,membership-method}
\alias{member}
\alias{member,}
\alias{member,membership-method}
\alias{membership-method}
\alias{show,membership-method}
\title{Method for membership classes}
\usage{
\S4method{is.na}{membership}(x)

\S4method{show}{membership}(object)

member(x)

\S4method{member}{membership}(x)

hard.label(x)

\S4method{hard.label}{membership}(x)
}
\arguments{
\item{x}{an object}

\item{object}{an object}
}
\description{
Method for membership classes
}

