% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{membership}
\alias{membership}
\title{Make an Partition Matrix}
\usage{
membership(member, K, n, RandomNumber = 0)
}
\arguments{
\item{member}{membership matrix}

\item{K}{number of cluster}

\item{n}{number of observation}

\item{RandomNumber}{random number/seed used}
}
\value{
membership object
}
\description{
Make an Partition Matrix
}
\section{Slots}{

\describe{
\item{\code{member}}{membership object}

\item{\code{hard.label}}{hard label}
}}
\examples{
#make an matrix membership
membership(K=3,n=20,RandomNumber=1234)
}

