% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minWeightBipartiteMatching.R
\name{minWeightBipartiteMatching}
\alias{minWeightBipartiteMatching}
\title{Matching label}
\usage{
minWeightBipartiteMatching(x, y)
}
\arguments{
\item{x}{an fuzzycluster object}

\item{y}{an fuzzy cluster object}
}
\description{
Matching label of 2 cluster via Hungary Algorithm
}

