% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_co_fuzzycluster.R
\name{print.co_fuzzycluster}
\alias{print.co_fuzzycluster}
\title{Print Consensus Fuzzy Clustering Result}
\usage{
\method{print}{co_fuzzycluster}(x, ...)
}
\arguments{
\item{x}{consensus fuzzy clustering object}

\item{...}{another paramater}
}
\description{
Print Consensus Fuzzy Clustering
}

