% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_Membership.R
\name{print.membership}
\alias{print.membership}
\title{Print membership}
\usage{
\method{print}{membership}(x, ...)
}
\arguments{
\item{x}{membership object}

\item{...}{another parameter}
}
\description{
Print membership object
}

