% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_validation.R
\name{print.validation}
\alias{print.validation}
\title{Print Validation Index}
\usage{
\method{print}{validation}(x, ...)
}
\arguments{
\item{x}{validation object}

\item{...}{another parameter}
}
\description{
Print Validation Index for fuzzy clustering
}

