% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarplot.R
\name{radar.plotting}
\alias{radar.plotting}
\title{Radar Ploting Fuzzy Cluster Result}
\usage{
radar.plotting(object, data.X)
}
\arguments{
\item{object}{a fuzzycluster object}

\item{data.X}{a matrix data}
}
\description{
Radar Ploting Fuzzy Cluster Result
}
\details{
Make Visualization Radar Ploting
}
\examples{
fuzzy.CM(iris[,1:4],K=3,m=2,max.iteration=100,threshold=1e-5,RandomNumber=1234)->cl
radar.plotting(cl,iris[,1:4])
}

