% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_validation.R
\docType{methods}
\name{show,validation-method}
\alias{CE}
\alias{CE,}
\alias{CE,validation-method}
\alias{Kwon}
\alias{Kwon,}
\alias{Kwon,validation-method}
\alias{MPC}
\alias{MPC,}
\alias{MPC,validation-method}
\alias{PC}
\alias{PC,}
\alias{PC,validation-method}
\alias{S}
\alias{S,}
\alias{S,validation-method}
\alias{Tang}
\alias{Tang,}
\alias{Tang,validation-method}
\alias{XB}
\alias{XB,}
\alias{XB,validation-method}
\alias{show,validation-method}
\alias{validation-method}
\title{Method for validation classes}
\usage{
\S4method{show}{validation}(object)

PC(x)

\S4method{PC}{validation}(x)

MPC(x)

\S4method{MPC}{validation}(x)

CE(x)

\S4method{CE}{validation}(x)

XB(x)

\S4method{XB}{validation}(x)

S(x)

\S4method{S}{validation}(x)

Tang(x)

\S4method{Tang}{validation}(x)

Kwon(x)

\S4method{Kwon}{validation}(x)
}
\arguments{
\item{object}{an object}

\item{x}{an object}
}
\description{
Method for validation classes
}

