% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_fuzzycluster.R
\name{c_fuzzycluster}
\alias{c_fuzzycluster}
\title{Combine fuzzy cluster result}
\usage{
c_fuzzycluster(x, ...)
}
\arguments{
\item{x}{a fuzzy cluster object}

\item{...}{a fuzzy cluster object}
}
\description{
combining fuzzy cluster result before ensembling
}

