% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_membership.R
\name{membership-class}
\alias{initMembership}
\alias{membership-class}
\title{Membership}
\usage{
initMembership()
}
\description{
Membership
}
\section{Slots}{

\describe{
\item{\code{member}}{membership matrix}

\item{\code{hard.label}}{vector of hard hard.labeling}
}}

