% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{membership}
\alias{membership}
\title{Make an Partition Matrix}
\usage{
membership(member, K, n, RandomNumber = 0)
}
\arguments{
\item{member}{membership matrix}

\item{K}{number of cluster}

\item{n}{number of observation}

\item{RandomNumber}{random number/seed used}
}
\description{
Make an Partition Matrix
}

