% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarplot.R
\name{radar.plotting}
\alias{radar.plotting}
\title{Radar Ploting Fuzzy Cluster Result}
\usage{
radar.plotting(object, data.X)
}
\arguments{
\item{object}{a fuzzycluster object}

\item{data.X}{a matrix data}
}
\description{
Radar Ploting Fuzzy Cluster Result
}
\details{
Make Visualization Radar Ploting
}

