% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_validation.R
\docType{class}
\name{validation-class}
\alias{validation-class}
\title{Validation Class}
\description{
Validation Class
}
\section{Slots}{

\describe{
\item{\code{PC}}{Partition Coefficient index}

\item{\code{MPC}}{Modified Partition Coefficient index}

\item{\code{CE}}{Classification Entropy index}

\item{\code{S}}{Separation index}

\item{\code{XB}}{Xie Beni index}

\item{\code{Kwon}}{Kwon index}

\item{\code{Tang}}{Tang index}
}}

